# optimized GetCountryByNameConsiderTranslation in CountryProviderExtension.cs