### Benchmarks


| Method          | CountryName | Mean      | Error    | StdDev   |	Execution Time Comparison |
|---------------- |------------ |----------:|---------:|---------:|--------------------------:|
| PreOptimization | Danmark     |  24.68 ms | 0.175 ms | 0.146 ms |   	  	+117479% 	     |
| PreOptimization | Denmark     |  25.15 ms | 0.480 ms | 0.658 ms |	  	  	+120062% 	     | 
| PreOptimization | India       |  42.49 ms | 0.592 ms | 0.525 ms |	  	  	+116727%       |
| PreOptimization | Ireland     |  47.11 ms | 0.792 ms | 0.778 ms |	  	  	+122774%       |
| PreOptimization | Norge       |  74.14 ms | 1.312 ms | 1.227 ms |   	  	+96233%        |
| PreOptimization | Norway      |  75.42 ms | 0.702 ms | 0.657 ms |   	  	+94010%        |
| PreOptimization | Sverige     | 100.58 ms | 1.421 ms | 1.187 ms |   	  	+82825%        |
| PreOptimization | Sweden      |  97.38 ms | 1.846 ms | 1.727 ms |   	  	+84740%        |

| Method           | CountryName | Mean      | Error    | StdDev   | Execution Time Comparison |
|----------------- |------------ |----------:|---------:|---------:|---------------------:|
| PostOptimization | Danmark     |  20.99 us | 0.411 us | 0.422 us |  	   -99.91%	      |
| PostOptimization | Denmark     |  20.93 us | 0.257 us | 0.241 us |  	   -99.91%	      |
| PostOptimization | India       |  36.37 us | 0.332 us | 0.310 us |  	   -99.91%	      |
| PostOptimization | Ireland     |  38.34 us | 0.268 us | 0.250 us |  	   -99.91%	      |
| PostOptimization | Norge       |  78.00 us | 0.869 us | 0.813 us |  	   -99.89%	      |
| PostOptimization | Norway      |  80.14 us | 1.295 us | 1.591 us |  	   -99.89%	      |
| PostOptimization | Sverige     | 117.43 us | 1.233 us | 1.154 us |  	   -99.88%	      |
| PostOptimization | Sweden      | 114.78 us | 1.028 us | 0.858 us |  	   -99.88%   	  |