@tinohager 2024-08-20
Optimize performance of GetCountryByNameConsiderTranslation
https://github.com/nager/Nager.Country/commit/e4c8350a3d374d671039a2ce81d37644ec9e0671